Linhchelsea

+++ app/Http/Controllers/BackEnd/CustomerReviewController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\BackEnd;
 
 use App\CustomerReview;
+use App\Http\Requests\ReviewRequest;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\File;
@@ -40,16 +41,17 @@ class CustomerReviewController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ReviewRequest $request)
     {
         $review = new CustomerReview();
         $review->name = $request->fullname;
         $review->content = $request->reviewContent;
-        $filename = "";
         if($request->file('image') != null){
             $image = $request->file('image')->store('public/reviews');
             $arr_filename = explode("/",$image);
             $filename = end($arr_filename);
+        }else{
+            $filename = 'avatar.png';
         }
         $review->image = $filename;
         if($review->save()){
@@ -79,14 +81,17 @@ class CustomerReviewController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(ReviewRequest $request, $id)
     {
         $review = CustomerReview::findOrFail($id);
         $review->name = $request->fullname;
         $review->content = $request->reviewContent;
         if($request->file('image') != null){
-            //Xoa anh cu~
-            File::delete('storage/reviews/'.$review->image);
+            if($review->image != 'avatar.png'){
+                //Xoa anh cu~
+                File::delete('storage/reviews/'.$review->image);
+            }
+
             $image = $request->file('image')->store('public/reviews');
             $arr_filename = explode("/",$image);
             $filename = end($arr_filename);
@@ -111,7 +116,10 @@ class CustomerReviewController extends Controller
     public function destroy(Request $request, $id)
     {
         $review = CustomerReview::findOrFail($id);
-        File::delete('storage/reviews/'.$review->image);
+        if($review->image != 'avatar.png'){
+            //Xoa anh cu~
+            File::delete('storage/reviews/'.$review->image);
+        }
         if($review->delete()){
             $request->session()->flash('success','Delete successfully!');
         }else{
