+++ app/Http/Controllers/BackEnd/GalleryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\BackEnd;
 
 use App\Gallery;
+use App\Http\Requests\GalleryRequest;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\File;
@@ -40,15 +41,16 @@ class GalleryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(GalleryRequest $request)
     {
         $gallery = new Gallery();
         $gallery->title = $request->title;
-        $filename = "";
         if($request->file('image') != null){
             $image = $request->file('image')->store('public/gallery');
             $arr_filename = explode("/",$image);
             $filename = end($arr_filename);
+        }else{
+            $filename = 'default.png';
         }
         $gallery->image = $filename;
         if($gallery->save()){
@@ -79,13 +81,15 @@ class GalleryController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(GalleryRequest $request, $id)
     {
         $gallery = Gallery::findOrFail($id);
         $gallery->title = $request->title;
         if($request->file('image') != null){
-            //Xoa anh cu~
-            File::delete('storage/gallery/'.$gallery->image);
+            if($gallery->image != 'default.png'){
+                //Xoa anh cu~
+                File::delete('storage/gallery/'.$gallery->image);
+            }
             $image = $request->file('image')->store('public/gallery');
             $arr_filename = explode("/",$image);
             $filename = end($arr_filename);
@@ -110,7 +114,10 @@ class GalleryController extends Controller
     public function destroy(Request $request,$id)
     {
         $gallery = Gallery::findOrFail($id);
-        File::delete('storage/gallery/'.$gallery->image);
+        if($gallery->image != 'default.png'){
+            //Xoa anh cu~
+            File::delete('storage/gallery/'.$gallery->image);
+        }
         if($gallery->delete()){
             $request->session()->flash('success','Delete successfully!');
         }else{
